Opens PTM Masterlist

- Pulls from the PTM Masterlist Google Sheet and reads the file - *note* you must have been granted access on your gmail before using

masterlist()

Value

dataframe

Examples

masterlist()

#> Using an auto-discovered, cached token.
#> To suppress this message, modify your code or options to clearly consent to the use of a cached token.
#> See gargle's "Non-interactive auth" vignette for more details:
#> https://gargle.r-lib.org/articles/non-interactive-auth.html

#> The googlesheets4 package is using a cached token for jasmine.lai.0911@gmail.com.

#> Auto-refreshing stale OAuth token.

#> Reading from "PTM - Master List"

#> Range "'All PTM data'!2:5000000"

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> # A tibble: 1,964 x 39
#>    `PTM#` `Final determin… `genetic match/… `Determination … Qualifier
#>     <int> <chr>            <chr>            <chr>            <chr>    
#>  1     NA Genus(capitaliz… separated by se… Genus(capitaliz… any unce…
#>  2      1 Corallina offic… NA               Corallina offic… NA       
#>  3      2 Bossiella plumo… no more DNA      Bossiella plumo… NA       
#>  4      3 Calliarthron ch… NA               Calliarthron ch… NA       
#>  5      4 Calliarthron tu… NA               Calliarthron tu… NA       
#>  6      5 Corallina vanco… NA               Corallina vanco… NA       
#>  7      6 Calliarthron tu… NA               Calliarthron tu… NA       
#>  8      7 Calliarthron ch… NA               Calliarthron ch… NA       
#>  9      8 Lithothrix aspe… NA               Lithothrix aspe… NA       
#> 10      9 Bossiella dicho… NA               Bossiella dicho… NA       
#> # … with 1,954 more rows, and 34 more variables: `DNA location` <chr>, `Extra
#> #   DNA Sample` <chr>, `Date of Extraction` <date>, `PCR results` <chr>,
#> #   CO1 <chr>, psbA <chr>, rbcL <chr>, `Voucher location` <chr>,
#> #   BOLDNumber <chr>, GenbankNumber <chr>, `Red/Coralline/Green/Brown` <chr>,
#> #   Country <chr>, StateProvince <chr>, Locality <chr>, `Date Collected` <chr>,
#> #   `Reproductive Status` <chr>, `Field Notes` <chr>, `Accession #` <chr>,
#> #   `Missing accession` <chr>, Habitat <chr>, Depth <chr>, `Fragments
#> #   available?` <chr>, Latitude <dbl>, Longitude <dbl>, `Primary
#> #   Collector` <chr>, `Other collectors` <chr>, `Determined by` <chr>,
#> #   `Determination Date` <chr>, Project <chr>, `Photos on server` <int>,
#> #   `Photos on server NEW test` <int>, `CO1 NEW` <chr>, `PSBA NEW` <chr>, `RBCL
#> #   NEW` <chr>