Searches for species in PTM masterlist looks in the Final determination and Determination in the field

Searches for species in PTM masterlist looks in the Final determination and Determination in the field

search_ptm(name)

Arguments

name
scientific name of seaweed

Value

dataframe

Examples

search_ptm("Ulva lactuca")

#> Reading from "PTM - Master List"

#> Range "'All PTM data'!2:5000000"

#> # A tibble: 0 x 39
#> # … with 39 variables: `PTM#` <int>, `Final determination` <chr>, `genetic
#> #   match/notes` <chr>, `Determination in the field` <chr>, Qualifier <chr>,
#> #   `DNA location` <chr>, `Extra DNA Sample` <chr>, `Date of
#> #   Extraction` <date>, `PCR results` <chr>, CO1 <chr>, psbA <chr>, rbcL <chr>,
#> #   `Voucher location` <chr>, BOLDNumber <chr>, GenbankNumber <chr>,
#> #   `Red/Coralline/Green/Brown` <chr>, Country <chr>, StateProvince <chr>,
#> #   Locality <chr>, `Date Collected` <chr>, `Reproductive Status` <chr>, `Field
#> #   Notes` <chr>, `Accession #` <chr>, `Missing accession` <chr>,
#> #   Habitat <chr>, Depth <chr>, `Fragments available?` <chr>, Latitude <dbl>,
#> #   Longitude <dbl>, `Primary Collector` <chr>, `Other collectors` <chr>,
#> #   `Determined by` <chr>, `Determination Date` <chr>, Project <chr>, `Photos
#> #   on server` <int>, `Photos on server NEW test` <int>, `CO1 NEW` <chr>, `PSBA
#> #   NEW` <chr>, `RBCL NEW` <chr>